Move all python files into namespaces
