+++ gaia/utils/utils.py
