+++ tests/test_gamestate.py
@@ -1,8 +1,8 @@
 import pytest
 
-from gaia.gamestate import ResearchBoard
-from gaia.enums import ResearchTracks
-from gaia.players import Income
+from gaia.gamestate.gamestate import ResearchBoard
+from gaia.utils.enums import ResearchTracks
+from gaia.gamestate.players import Income
 from tests.util import get_research_bonus_func_for_track
