+++ tests/test_player.py
@@ -1,5 +1,5 @@
 import pytest
-from gaia.players import PlayerResources
+from gaia.gamestate.players import PlayerResources
 
 
 @pytest.mark.parametrize("power_before, gain_amount, power_after, description", [
